issue/282 - restructured oom exception handling
